historical-data-migration/requirements.txt
pylint
pytest
pymssql
boto3
python-dotenv
pandas